HADOOP_11203: Allow distcp to accept the max bandwidth value in fraction of MBs

+++ src/main/java/org/apache/hadoop/tools/DistCpOptions.java
@@ -54,7 +54,7 @@ public class DistCpOptions {
   }
 
   private int maxMaps = DistCpConstants.DEFAULT_MAPS;
-  private int mapBandwidth = DistCpConstants.DEFAULT_BANDWIDTH_MB;
+  private float mapBandwidth = DistCpConstants.DEFAULT_BANDWIDTH_MB;
 
   private String sslConfigurationFile;
 
@@ -293,7 +293,7 @@ public class DistCpOptions {
    *
    * @return Bandwidth in MB
    */
-  public int getMapBandwidth() {
+  public float getMapBandwidth() {
     return mapBandwidth;
   }
 
@@ -302,7 +302,7 @@ public class DistCpOptions {
    *
    * @param mapBandwidth - per map bandwidth
    */
-  public void setMapBandwidth(int mapBandwidth) {
+  public void setMapBandwidth(float mapBandwidth) {
     if (mapBandwidth <= 0) {
       throw new IllegalArgumentException("Bandwidth " +
           mapBandwidth + " is invalid (should be > 0)");
