+++ src/main/java/org/apache/hadoop/tools/util/ThrottledInputStream.java
@@ -24,7 +24,7 @@ import java.io.InputStream;
 public class ThrottledInputStream extends InputStream {
 
   private final InputStream rawStream;
-  private final long maxBytesPerSec;
+  private final float maxBytesPerSec;
   private final long startTime = System.currentTimeMillis();
 
   private long bytesRead = 0;
@@ -36,7 +36,7 @@ public class ThrottledInputStream extends InputStream {
     this(rawStream, Long.MAX_VALUE);
   }
 
-  public ThrottledInputStream(InputStream rawStream, long maxBytesPerSec) {
+  public ThrottledInputStream(InputStream rawStream, float maxBytesPerSec) {
     assert maxBytesPerSec > 0 : "Bandwidth " + maxBytesPerSec + " is invalid"; 
     this.rawStream = rawStream;
     this.maxBytesPerSec = maxBytesPerSec;
